Move compare_version_strings to string_utils.h
