+++ test/apprun/runtime_interpreter_test.c
@@ -30,6 +30,7 @@
 
 #include "apprun/runtime_interpreter.h"
 #include "../common/tests_shared.h"
+#include "common/string_utils.h"
 
 void test_parse_ld_trace_lib_path() {
     printf("%s: ", __PRETTY_FUNCTION__);
